spec parsing has been moved into separated process
