+++ core/file-tree/data-collector.js
@@ -1,0 +1,20 @@
+var path = require('path');
+var parser = require(path.join(__dirname, 'specs-parser'));
+
+process.on('message', function(msg) {
+
+	this.dataCollectionTask = function(message) {
+		parser.init(message.config);
+		parser.parse(message.config.specsRoot, function(specsData) {
+			process.send(specsData);
+		});
+	};
+	
+	this._init = function() {
+		if(msg) {
+			this.dataCollectionTask(msg);
+		} else {
+			console.log("Child process resieved empty message. Unable to start processing.");
+		}
+	}.bind(this)();
+});
